+++ src/types/literal.rs
@@ -25,3 +25,31 @@ impl<'a, T> Literal<'a, T> {
         matches!(self, Self::NotAtomicFormula(..))
     }
 }
+
+impl<'a, T> From<AtomicFormula<'a, T>> for Literal<'a, T> {
+    fn from(value: AtomicFormula<'a, T>) -> Self {
+        Literal::new(value)
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::parsers::{atomic_formula, parse_term};
+    use crate::Term;
+
+    #[test]
+    fn from_works() {
+        let input = "(= x y)";
+        let (_, effect) = atomic_formula(parse_term)(input).unwrap();
+
+        let literal: Literal<Term> = effect.into();
+        assert_eq!(
+            literal,
+            Literal::AtomicFormula(AtomicFormula::new_equality(
+                Term::new_name("x".into()),
+                Term::new_name("y".into()),
+            ))
+        )
+    }
+}